+++ module03/ex02/ClapTrap.hpp
@@ -67,7 +67,7 @@ class  ClapTrap
   // Attributs
   
   protected :
-  	int HitPoints;
+  	long int HitPoints;
   	int MaxHitPoints;
   	int EnergyPoints;
   	int MaxEnergyPoints;
